refactor: pages/ 폴더는 오직 라우팅 및 페이지 단위 조합으로만 수행 (#5)
